requirements.txt
Django==1.8
Padding==0.4
dj-database-url==0.3.0
django-braces==1.4.0
django-clear-cache==0.3
django-cors-headers==1.0.0
django-heroku-memcacheify==0.8
django-jsonfield==0.9.13
django-markdown-deux==1.0.5
django-model-utils==2.2
django-pylibmc==0.5.0
djangorestframework==3.1.0
drf-json-api==0.1.0
gunicorn==19.1.1
markdown2==2.3.0
psycopg2==2.7.3.2
pycrypto==2.6.1
pylibmc==1.4.1
requests==2.5.1
six==1.8.0
whitenoise==1.0.3
wsgiref==0.1.2
django-url-filter==0.3.15

